Add new file for filter functions
